requirements.txt
django
python-dotenv
django-debug-toolbar
django_filter